server/requirements.txt
fastapi==0.111.0
python-socketio[asyncio]==5.11.0
uvicorn==0.29.0
pydantic==2.7.4
openai==1.30.5
python-dotenv